Done
There is a bug where the spaces in the words do not appear correctly. For example, for the word "being helpful". After showing "being",&nbsp;the space should make the right border/cursor move a little bit to the right (to account for the space being entered by the user) yet it doesn't work correctly. Please fix this error.